+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `ContractCreateFlow.[set|get]AutoRenewAccountId()`
  - Client now automatically updates the network via a mirror node query at regular intervals.  You can set/get the interval with `Client.[set|get]NetworkUpdatePeriod()`
  - Client can now be set from a `NodeAddressBook` with `Client.setNetworkFromAddressBook()`
+ - `Client.setMirrorTransportSecurity()`
+ - `Client.mirrorIsTransportSecurity()`
 
 ### Fixed
 
@@ -22,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `validateChecksum()`, `toStringWithChecksum()`, `hashCode()`, `equals()`, and `compareTo()` now function correctly for `AccountId`s with `aliasEvmAddress`es.
  - Changed the default transaction fee for `AccountCreateTransaction` to 5 Hbar.
  - `PrivateKey.is[ED25519|ECDSA]()` is now correct for ED25519 private keys.
+ - Default mirror node for mainnet is now `mainnet-public.mirrornode.hedera.com:443` instead of `hcs.mainnet.mirrornode.hedera.com:5600`
 
 ## v2.17.3